requirements.txt
absl-py==0.8.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
future==0.17.1
grpcio==1.16.1
idna==2.8
jsonpatch==1.24
jsonpointer==2.0
Markdown==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.2
olefile==0.46
Pillow==6.2.0
protobuf==3.9.2
pycparser==2.19
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
scipy==1.3.1
six==1.12.0
tensorboard==2.0.0
pytorch==1.3.0
torchfile==0.1.0
torchvision==0.4.1a0+d94043a
tornado==6.0.3
tqdm==4.36.1
urllib3==1.25.6
visdom==0.1.8.9
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
